app/__init__.py
import os

from flask import Flask
from flask_jwt_extended import (
    JWTManager,
)
from flask_migrate import Migrate
from flask_sqlalchemy import SQLAlchemy
from flask_marshmallow import Marshmallow

from dotenv import load_dotenv




app = Flask(__name__)

# = mysql+pymysql://usuario:contraseña@ip/nombre_db
app.config['SQLALCHEMY_DATABASE_URI'] = os.environ['SQLALCHEMY_DATABASE_URI']
app.config['JWT_SECRET_KEY'] = os.environ['JWT_SECRET_KEY']

db = SQLAlchemy(app=app)
migrate = Migrate(app, db)
jwt = JWTManager(app)
ma = Marshmallow(app)
load_dotenv()

from app.views import view

